Open step folder
