+++ pymetaheuristics/genetic_algorithm/steps/crossovers.py
@@ -1,25 +1,8 @@
+from random import randint
 from typing import Tuple
-from random import choices, randint, randrange, random
 
+from pymetaheuristics.genetic_algorithm.types import Genome
 from pymetaheuristics.genetic_algorithm.exceptions import CrossOverException
-from pymetaheuristics.genetic_algorithm.types import (
-    FitnessFunction, Genome, Population)
-
-
-def random_weighted_selection(
-    population: Population,
-    fitness_function: FitnessFunction,
-    k: int = 2
-) -> Population:
-    """Selects randomly k genomes on a population. This approach considers the
-    fitness of each Genome as weights so the most fitted is very likely to be
-    choosen, but, still gives room for a little of jumps.
-    """
-    return choices(
-        population=population,
-        weights=[fitness_function(genome) for genome in population],
-        k=k
-    )
 
 
 def single_point_crossover(g1: Genome, g2: Genome) -> Tuple[Genome, Genome]:
@@ -76,20 +59,3 @@ def pmx_single_point(g1: Genome, g2: Genome) -> Tuple[Genome, Genome]:
         g2child[i] = g1[i]
 
     return g1child, g2child
-
-
-def inter_mutation(
-    genome: Genome,
-    q: int = 2,
-    probability: float = 0.75
-) -> Genome:
-    """At a random chance, change interposition of q genes on the Genome."""
-    for _ in range(q):
-        index = randrange(len(genome))
-
-        if random() > probability:
-            return genome
-
-        genome[index], genome[index - 1] = genome[index - 1], genome[index]
-
-    return genome
